requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.25.0

langchain-core==0.3.66
# langchain-community==0.3.26
langchain-experimental==0.3.4
langchain-text-splitters==0.3.8

llama-index==0.12.43
llama-index-llms-langchain==0.6.0
llama-index-vector-stores-milvus==0.8.5
llama-index-embeddings-huggingface==0.5.4
llama-index-readers-file==0.4.9

pymilvus==2.5.11
pymssql==2.3.1
sqlalchemy==2.0.30

sentence-transformers>=2.2.2,<3.2.0   # 避免最新版可能不兼容
pypdf==5.6.0
python-dotenv==1.0.1

+ langchain-ollama==0.2.1      # 新增 Ollama 整合
+ ollama==0.4.3 